[Refactor] #12 서비스를 인터페이스와 구현체로 분리
